# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ProofGoal API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /insurance/products:
    get:
      operationId: listInsuranceProducts
      tags: [insurance]
      summary: List available insurance products
      parameters:
        - name: type
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/InsuranceType"
      responses:
        "200":
          description: List of insurance products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InsuranceProduct"
    post:
      operationId: createInsuranceProduct
      tags: [insurance]
      summary: Create a user-defined insurance product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsuranceProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsuranceProduct"
  /insurance/products/{productId}:
    get:
      operationId: getInsuranceProduct
      tags: [insurance]
      summary: Get insurance product detail
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
      responses:
        "200":
          description: Insurance product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsuranceProduct"
        "404":
          description: Not found
  /insurance/policies:
    get:
      operationId: listInsurancePolicies
      tags: [insurance]
      summary: List insurance policies for a wallet
      parameters:
        - $ref: "#/components/parameters/WalletAddressQueryParam"
      responses:
        "200":
          description: List of policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InsurancePolicy"
    post:
      operationId: purchaseInsurancePolicy
      tags: [insurance]
      summary: Purchase an insurance policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsurancePolicyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsurancePolicy"
  /insurance/policies/{policyId}:
    get:
      operationId: getInsurancePolicy
      tags: [insurance]
      summary: Get insurance policy detail
      parameters:
        - $ref: "#/components/parameters/PolicyIdParam"
      responses:
        "200":
          description: Policy detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsurancePolicy"
        "404":
          description: Not found
...
  /governance/proposals:
    get:
      operationId: listGovernanceProposals
      tags: [governance]
      summary: List governance proposals
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ProposalStatus"
      responses:
        "200":
          description: List of proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GovernanceProposal"
    post:
      operationId: createGovernanceProposal
      tags: [governance]
      summary: Create a new governance proposal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GovernanceProposalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceProposal"
  /governance/proposals/{proposalId}:
    get:
      operationId: getGovernanceProposal
      tags: [governance]
      summary: Get governance proposal detail with vote tallies
      parameters:
        - $ref: "#/components/parameters/ProposalIdParam"
      responses:
        "200":
          description: Proposal detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceProposal"
        "404":
          description: Not found
  /governance/proposals/{proposalId}/votes:
    post:
      operationId: castVote
      tags: [governance]
      summary: Cast a vote on a governance proposal
      parameters:
        - $ref: "#/components/parameters/ProposalIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteInput"
      responses:
        "201":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vote"
components:
  parameters:
...
    WalletAddressQueryParam:
      name: walletAddress
      in: query
      required: true
      schema:
        type: string
...
    ProductIdParam:
      name: productId
      in: path
      required: true
      schema:
        type: string
    PolicyIdParam:
      name: policyId
      in: path
      required: true
      schema:
        type: string
...
    ProposalIdParam:
      name: proposalId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    InsuranceType:
      type: string
      enum:
        [
          favorite_team_loss,
          tournament_exit,
          qualification,
          goal_insurance,
          event_triggered,
          custom,
        ]
    InsuranceProduct:
      type: object
      required:
        [
          id,
          name,
          type,
          description,
          premiumRateBps,
          maxCoverageLamports,
          triggerCondition,
          createdAt,
        ]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          $ref: "#/components/schemas/InsuranceType"
        description:
          type: string
        premiumRateBps:
          type: integer
        maxCoverageLamports:
          type: number
        triggerCondition:
          type: string
        createdAt:
          type: string
          format: date-time
    InsuranceProductInput:
      type: object
      required:
        [name, type, description, premiumRateBps, maxCoverageLamports, triggerCondition]
      properties:
        name:
          type: string
          minLength: 3
        type:
          $ref: "#/components/schemas/InsuranceType"
        description:
          type: string
        premiumRateBps:
          type: integer
          minimum: 1
          maximum: 5000
        maxCoverageLamports:
          type: number
          minimum: 1
        triggerCondition:
          type: string
    PolicyStatus:
      type: string
      enum: [active, triggered, expired, claimed, void]
    InsurancePolicy:
      type: object
      required:
        [
          id,
          walletAddress,
          productId,
          matchId,
          premiumPaidLamports,
          coverageLamports,
          status,
          purchasedAt,
        ]
      properties:
        id:
          type: string
        walletAddress:
          type: string
        productId:
          type: string
        matchId:
          type: ["string", "null"]
        premiumPaidLamports:
          type: number
        coverageLamports:
          type: number
        status:
          $ref: "#/components/schemas/PolicyStatus"
        purchasedAt:
          type: string
          format: date-time
        claimTxSig:
          type: ["string", "null"]
    InsurancePolicyInput:
      type: object
      required: [walletAddress, productId, coverageLamports]
      properties:
        walletAddress:
          type: string
        productId:
          type: string
        matchId:
          type: ["string", "null"]
        coverageLamports:
          type: number
          minimum: 1
...
    ProposalStatus:
      type: string
      enum: [active, passed, rejected, expired]
    GovernanceProposal:
      type: object
      required:
        [
          id,
          title,
          description,
          status,
          votesFor,
          votesAgainst,
          createdAt,
          endsAt,
        ]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/ProposalStatus"
        votesFor:
          type: number
        votesAgainst:
          type: number
        createdAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
    GovernanceProposalInput:
      type: object
      required: [title, description, endsAt]
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
          minLength: 3
        endsAt:
          type: string
          format: date-time
    VoteChoice:
      type: string
      enum: [for, against]
    VoteInput:
      type: object
      required: [walletAddress, choice, weight]
      properties:
        walletAddress:
          type: string
        choice:
          $ref: "#/components/schemas/VoteChoice"
        weight:
          type: number
          minimum: 0.000000001
    Vote:
      type: object
      required: [id, proposalId, walletAddress, choice, weight, createdAt]
      properties:
        id:
          type: string
        proposalId:
          type: string
        walletAddress:
          type: string
        choice:
          $ref: "#/components/schemas/VoteChoice"
        weight:
          type: number
        createdAt:
          type: string
          format: date-time