# lib/api-zod/src/generated/types/insuranceType.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * ProofGoal API specification
 * OpenAPI spec version: 0.1.0
 */

export type InsuranceType = typeof InsuranceType[keyof typeof InsuranceType];


export const InsuranceType = {
  favorite_team_loss: 'favorite_team_loss',
  tournament_exit: 'tournament_exit',
  qualification: 'qualification',
  goal_insurance: 'goal_insurance',
  event_triggered: 'event_triggered',
  custom: 'custom',
} as const;